server/requirements.txt
asgiref==3.3.1
cffi==1.14.4
cryptography==3.3.1
Django==3.1.3
django-rest-auth==0.9.5
django-rest-knox==4.1.0
djangorestframework==3.12.2
gunicorn==20.0.4
paho-mqtt==1.5.1
pycparser==2.20
python-decouple==3.3
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
whitenoise==5.2.0
